feat: add user authentication to backend

+++ backend/src/services/recipes.js
@@ -1,8 +1,8 @@
 import { Recipe } from '../db/models/recipe.js'
 
 // Create recipe
-export async function createRecipe({ title, author, contents, tags, image }) {
-  const recipe = new Recipe({ title, author, contents, tags, image })
+export async function createRecipe(userId, { title, contents, tags, image }) {
+  const recipe = new Recipe({ title, author: userId, contents, tags, image })
   return await recipe.save()
 }
 
@@ -35,17 +35,18 @@ export async function getRecipeById(recipeId) {
 
 // Update recipe
 export async function updateRecipe(
+  userId,
   recipeId,
-  { title, author, contents, tags, image },
+  { title, contents, tags, image },
 ) {
   return await Recipe.findOneAndUpdate(
     { _id: recipeId },
-    { $set: { title, author, contents, tags, image } },
+    { $set: { title, author: userId, contents, tags, image } },
     { new: true },
   )
 }
 
 // Delete recipe
-export async function deleteRecipe(recipeId) {
-  return await Recipe.deleteOne({ _id: recipeId })
+export async function deleteRecipe(userId, recipeId) {
+  return await Recipe.deleteOne({ _id: recipeId, author: userId })
 }
